+++ include/leanstore/concurrency/WalPayloadHandler.hpp
@@ -1,19 +1,17 @@
 #pragma once
 
-#include "leanstore/Exceptions.hpp"
 #include "leanstore/concurrency/GroupCommitter.hpp"
 #include "leanstore/concurrency/Worker.hpp"
-#include "leanstore/utils/Defer.hpp"
 
 namespace leanstore::cr {
 
 template <typename T>
 class WalPayloadHandler {
 public:
-  // payload of the active WAL
+  //! payload of the active WAL
   T* mWalPayload;
 
-  // size of the whole WalEntry, including payloads
+  //! size of the whole WalEntry, including payloads
   uint64_t mTotalSize;
 
 public:
@@ -39,20 +37,6 @@ public:
 
 template <typename T>
 inline void WalPayloadHandler<T>::SubmitWal() {
-  SCOPED_DEFER(DEBUG_BLOCK(){
-      // rapidjson::Document walDoc(rapidjson::kObjectType);
-      // auto* walEntry = cr::Worker::My().mLogging.mActiveWALEntryComplex;
-      // WalEntry::ToJson(walEntry, &walDoc);
-
-      // rapidjson::StringBuffer buffer;
-      // rapidjson::Writer<rapidjson::StringBuffer> writer(buffer);
-      // walDoc.Accept(writer);
-      // LS_DLOG("SubmitWal, workerId={}, startTs={}, curGsn={}, walJson={}",
-      //            Worker::My().mWorkerId, Worker::My().mActiveTx.mStartTs,
-      //            Worker::My().mLogging.GetCurrentGsn(),
-      //            WalEntry::ToJsonString(mWalPayload));
-  });
-
   cr::Worker::My().mLogging.SubmitWALEntryComplex(mTotalSize);
 }
